* modules/filters/mod_proxy_html.c:  Fix proxy_html_conf.bufsz   to have correct type, as it is used with ap_set_int_slot.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1876823 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/filters/mod_proxy_html.c
@@ -107,7 +107,7 @@ typedef struct {
     const char *doctype;
     const char *etag;
     unsigned int flags;
-    size_t bufsz;
+    int bufsz;
     apr_hash_t *links;
     apr_array_header_t *events;
     const char *charset_out;
